Add Create new Game object Topic
This is an attempt to add an MQTT topic allowing agents from Gama to create game objects on a unity scene. In this version, supported game object are Capsule, Cube, Cylinder, Plane, and Sphere.
It is possible to create a game object and specify its initial color. Support of setting up its initial position is not completely implemented yet.
Yet still, need to better manager MQTT message and avoid several game objects creation as is the case till now (?!).