Add Notificartion Topic
The idea is to introduce a generic mechanism allowing an agent to subscribe to a subject (variable) notification, managed by a notification manager on unity to be informed each time the subject in question is realized.  Example: Monitor the value of a variable in a unity object, and as soon as it exceeds a certain threshold, a notification will be sent to the subscribed agent.